Fix some issues with MNT
Don't use /mnt directly as the temporary mount point, as that's used in
real life by lots of things (especially in Chrome OS, where it's used in
things like the chroot).

Also unmount at the end of updatekernel.sh, and use $MNT in a couple
places where /mnt was accidentally hardcoded.