Add update_crosbuild_kernel.sh
Add a script that can be used if you're already comfortable building a
Chrome OS kernel in a Chrome OS chroot development environment. This can
be used if you've built a kernel for a specific board, and you want to
apply it on your USB stick. Note that you can't just compile the stock
Chrome OS kernel as-is and get it to work, it needs some changes like
adding fbcon and relaxing some of the LSM security features. I'm also
including the changes I made to 5.4 to get this to work on dedede.